--- a/crypto/groestl.h
+++ b/crypto/groestl.h
@@ -2,7 +2,7 @@
  * Trezor adaptation by Yura Pakhuchiy <pakhuchiy@gmail.com>. */
 /**
  * Groestl interface. This code implements Groestl with the recommended
- * parameters for SHA-3, with outputs of 224, 256, 384 and 512 bits.
+ * parameters for SHA-3, with output of 512 bits.
  *
  * ==========================(LICENSE BEGIN)============================
  *
@@ -29,7 +29,7 @@
  *
  * ===========================(LICENSE END)=============================
  *
- * @file     sph_groestl.h
+ * @file     groestl.h
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
 
@@ -39,7 +39,7 @@
 #include <stddef.h>
 
 /**
- * This structure is a context for Groestl-384 and Groestl-512 computations:
+ * This structure is a context for Groestl-512 computation:
  * it contains the intermediate values and some data from the last
  * entered block. Once a Groestl computation has been performed, the
  * context can be reused for another computation.