Solves https://github.com/trezor/trezor-firmware/issues/4079.

In this pull request, I optimized the implementations of blak256, blake2s, blake2b, groestl and AES in terms of flash size.

|       Name | Size before | Size after | Throughput before | Throughput after | Usage                                                                  |
|-----------:|------------:|-----------:|------------------:|-----------------:|------------------------------------------------------------------------|
|   blake256 |      2329 B |     1938 B |          1.5 MB/s |        0.59 MB/s | decred                                                                 |
|    blake2b |     15704 B |     2336 B |          2.3 MB/s |         1.2 MB/s | tezos, zcash, cardano                                                  |
|    blake2s |      4984 B |     1804 B |          3.1 MB/s |        0.96 MB/s | [firmware hash](https://github.com/trezor/trezor-firmware/blob/main/core/embed/extmod/modtrezorutils/modtrezorutils.c#L156), [bootloader check](https://github.com/trezor/trezor-firmware/blob/main/core/embed/firmware/bl_check.c#L152), [vendor header check](https://github.com/trezor/trezor-firmware/blob/main/core/embed/lib/image.c#L193) |in firmware |
| groestl512 |      9404 B |     6088 B |         0.16 MB/s |        0.13 MB/s | groestlcoin                                                            |
|        AES |     39964 B |     9538 B |                   |                  |                                                                        |
| AES128 ECB |             |            |         0.94 MB/s |        0.80 MB/s |                                                                        |
| AES128 CCM |             |            |        0.46 MB/s  |        0.43 MB/s | optiga secure channel                                                  |
| AES128 GCM |             |            |         0.31 MB/s |        0.30 MB/s | probably in THP                                                        |
|      ripemd160  |   5788 B | 1792 B | 5.5 MB/s | 1.3 MB/s | bitcoin, eos, ripple, nem |

 It's strange that it is possible to reduce the size of AES by half with only minimal impact on its speed.

There are still two places that would make sense to optimize, but I haven't addressed them yet:
  *  groestl512 that uses precomputed tables of size 4 KB
  *  ripemd160 that has 5 KB

EDITED: I fixed the "Size before" of AES, which should have been 20480 B higher.