@onvej-sl 2024-08-16
My suggestion is 
  * to use the memory-optimized versions of blake2b, AES and groestl512
  * keep the existing (speed-optimized) version of blake2s
 
I don't have a strong opinion on blake256, as it only saves a few hundred bytes. I'm not sure if it is worth diverging from the original implementation.

What do you think @andrewkozlik , @matejcik ?