@TychoVrahe 2024-08-16
+1 for keeping blake2s speed optimized, at least on F4 based models.