@matejcik 2024-08-23
> My suggestion is
> 
> * to use the memory-optimized versions of blake2b, AES and groestl512
> * keep the existing (speed-optimized) version of blake2s

agreed

> I don't have a strong opinion on blake256, as it only saves a few hundred bytes. I'm not sure if it is worth diverging from the original implementation.

let's keep the original then.

please go ahead and modify the PR accordingly.