@onvej-sl 2024-08-23
There are two things I'm considering:
1. In `groestl.c` and `blake2b.c`, I'm using `#ifdef` directives to switch between the original speed-optimized implementation and the new memory-optimized implementation. I'm considering whether to keep the `#ifdef` directives or to use the speed-optimized implementation directly.
2. `blake2b.c` and `groestl.c` are currently excluded from style formatting. However, I don't see a reason for this. The upstream implementations haven't changed for more than eight years. On the contrary, we occasionally make modifications to these implementations. Therefore, I'm considering including them in style formatting.

Any opinion, @andrewkozlik ?