@andrewkozlik 2024-08-27
1. I am inclined to keeping the `#ifdef` directives, including for blake2s, e.g. say we are running out of room in the boarloader and want to use the smaller variant of blake2s, then it will come in handy.

2.
    > On the contrary, we occasionally make modifications to these implementations.
    
    Until now our modifications were only variable initialization and `memzero()`. However, the changes in this PR are significant, making it difficult to keep in sync with upstream anyway if there were changes. So I agree about applying formatting. I'd consider doing it as a single commit in https://github.com/trezor/trezor-firmware/pull/4073 to keep the number of reformatting commits low.