@onvej-sl 2024-08-28
> Any update on ripemd160?

I didn't succeed in optimizing it myself. However a found [this](https://github.com/sipa/Coin25519/blob/master/src/crypto/ripemd160.c) implementation which is 1792 B in size and has decent speed. I added all the values to the table above.