@andrewkozlik 2024-09-02
As I mentioned earlier, I think it would be good have the `#ifdef` for blake2s as well. Firstly, it would be good to keep the blake2b and blake2s implementations consistent. Secondly, in the bootloader we may need to use the smaller blake2s version in the future.

That means we will need to add `OPTIMIZE_SIZE_BLAKE2S`, `OPTIMIZE_SIZE_BLAKE2B`, `OPTIMIZE_SIZE_ED25519`, etc., which is a good thing, because it explicitly documents what we are doing, whereas right now it is not obvious that blake2s is not size-optimized. Furthermore it documents what speed/size tradeoffs are available without having to investigate the source code. The definition in `options.h` should default to `OPTIMIZE_SIZE` when undefined to make it easy to use:
```
#ifndef OPTIMIZE_SIZE_BLAKE2S
#define OPTIMIZE_SIZE_BLAKE2S OPTIMIZE_SIZE
#endif

#ifndef OPTIMIZE_SIZE_BLAKE2B
#define OPTIMIZE_SIZE_BLAKE2B OPTIMIZE_SIZE
#endif

// ...
```