@onvej-sl 2024-09-09
@prusnak Your suggestion was addressed in https://github.com/trezor/trezor-firmware/pull/4107/commits/d707e54333d90b514548fa8c26e375c307d548f7. I think I cannot merge until you confirm it was addressed.